+++ Plugin.cs
@@ -3,32 +3,32 @@ using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
 
-namespace SolarExpanse.UIFramework
+namespace SolarExpanse.WindowManager
 {
     [BepInPlugin(PluginGuid, PluginName, PluginVersion)]
-    public class UiFrameworkPlugin : BaseUnityPlugin
+    public class WindowManagerPlugin : BaseUnityPlugin
     {
-        public const string PluginGuid = "com.mod.solarexpanse.uiframework";
-        public const string PluginName = "Solar Expanse UI Framework";
+        public const string PluginGuid = "com.mod.solarexpanse.windowmanager";
+        public const string PluginName = "Solar Expanse Window Manager";
         public const string PluginVersion = "1.4.0";
 
-        internal static ManualLogSource FrameworkLog;
+        internal static ManualLogSource WindowManagerLog;
 
         private Harmony _harmony;
 
         private void Awake()
         {
-            FrameworkLog = Logger;
-            SolarExpanseUi.SetLog(Logger);
+            WindowManagerLog = Logger;
+            SolarExpanseWindowManager.SetLog(Logger);
 
             _harmony = new Harmony(PluginGuid);
             _harmony.PatchAll();
             PauseScreenEscPatch.Apply(_harmony, Logger);
 
-            Logger.LogInfo("Solar Expanse UI Framework loaded");
+            Logger.LogInfo("Solar Expanse Window Manager loaded");
         }
 
-        private void Update() => SolarExpanseUi.InternalUpdate();
+        private void Update() => SolarExpanseWindowManager.InternalUpdate();
 
         private void LateUpdate() => PauseScreenEscPatch.LateUpdateTick();
     }